move common hdf5 code to a utility module
